secure Teacher endpoints in centralized config

+++ src/main/java/com/cisowski/schoolmanagement/grade/model/gradeType/GradeTypeResponse.java
@@ -2,11 +2,10 @@ package com.cisowski.schoolmanagement.grade.model.gradeType;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 
 @Data
 public class GradeTypeResponse {
-    private BigInteger id;
+    private Long id;
     private String gradeScope;
     private Double weight;
