+++ src/main/java/com/cisowski/schoolmanagement/grade/service/GradeTypeService.java
@@ -5,14 +5,13 @@ import com.cisowski.schoolmanagement.grade.model.gradeType.GradeTypeEntity;
 import com.cisowski.schoolmanagement.grade.model.gradeType.GradeTypeResponse;
 import com.cisowski.schoolmanagement.grade.model.gradeType.PatchGradeTypeRequest;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 public interface GradeTypeService {
     GradeTypeResponse addGradeType(AddGradeTypeRequest request);
-    GradeTypeResponse patchGradeType(PatchGradeTypeRequest request, BigInteger gradeTypeId);
-    void deleteGradeType(BigInteger gradeTypeId);
-    GradeTypeResponse getGradeType(BigInteger gradeTypeId);
+    GradeTypeResponse patchGradeType(PatchGradeTypeRequest request, Long gradeTypeId);
+    void deleteGradeType(Long gradeTypeId);
+    GradeTypeResponse getGradeType(Long gradeTypeId);
     Collection<GradeTypeResponse> getGradeTypes();
-    GradeTypeEntity fetchGradeType(BigInteger gradeTypeId);
+    GradeTypeEntity fetchGradeType(Long gradeTypeId);
 }
